Added LoginPage

+++ ExpenseTracker/ExpenseTracker/Model/Expenses.cs
@@ -20,7 +20,7 @@ namespace ExpenseTracker.Model
     {
         public int ExpenseId { get; set; }
         public string Description { get; set; }
-        public float ExpenseAmount { get; set; }
+        public decimal ExpenseAmount { get; set; }
         public DateTime ExpenseDate { get; set; }
         public Category ExpenseCategory { get; set; }
     }
